% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot.R
\name{xplot}
\alias{xplot}
\title{xplot}
\usage{
xplot(locations, localPath = getwd())
}
\arguments{
\item{locations}{A data.frame with location information.}

\item{localPath}{Path in the user's machine to download the images., Default: getwd()}
}
\value{
A ggplot object and an exported .jpeg image with the locations map.
}
\description{
Identifies the images from which locations must be accessed.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{gather}},\code{\link[tidyr]{pivot_wider}},\code{\link[tidyr]{nest}},\code{\link[tidyr]{c("nest", "nest", "nest")}},\code{\link[tidyr]{separate}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}},\code{\link[dplyr]{select}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{ranking}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{mutate_all}}
\code{\link[purrr]{map}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_polygon}},\code{\link[ggplot2]{map_data}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_label}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{coord_map}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{ggsave}}
}
