% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPolaris-package.R
\docType{package}
\name{XPolaris-package}
\alias{XPolaris}
\alias{XPolaris-package}
\title{XPolaris: Retrieving soil data from POLARIS}
\description{
The POLARIS database is a 30-meter probabilistic soil series map 
  of the contiguous United States (US). It represents an optimization of the Soil 
  Survey Geographic (SSURGO) database, circumventing issues of spatial disaggregation, 
  harmonizing, and filling spatial gaps. This R package facilitate the access to large 
  amounts of soil data within the US, currently stored as raster images.
}
\author{
\strong{Maintainer}: Luiz Moro Rosso \email{lhmrosso@ksu.edu}

Authors:
\itemize{
  \item Andre Borja Reis \email{andrefbr@ksu.edu}
  \item Adrian Correndo \email{correndo@ksu.edu}
  \item Ignacio Ciampitti \email{ciampitti@ksu.edu}
}

}
\keyword{internal}
