% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsoil.R
\name{xsoil}
\alias{xsoil}
\title{xsoil}
\usage{
xsoil(ximages_output, localPath = getwd())
}
\arguments{
\item{ximages_output}{PARAM_DESCRIPTION}

\item{localPath}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{case_when}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[sf]{st_as_sf}}
\code{\link[raster]{raster}},\code{\link[raster]{extract}}
\code{\link[tidyr]{nest}},\code{\link[tidyr]{pivot_wider}}
}
