% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exkansas}
\alias{exkansas}
\title{Example locations in Kansas}
\format{
This \code{data.frame} has 3 rows and the following 3 columns:
\describe{
\item{ID}{Unique identifier of a given location.}
\item{lat}{Latitude coordinate, in decimal degrees.}
\item{long}{Longitude coordinate, in decimal degrees.}
}
}
\usage{
exkansas
}
\description{
This locations example was created to help users setup their
own location data. The locations Scandia and Belleville are close in space,
to illustrate that only a minimum number of images will be downloaded to
get the soil information (proximal locations may fall in the same raster).
}
\keyword{datasets}
