% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ximages.R
\name{ximages}
\alias{ximages}
\title{ximages}
\usage{
ximages(locations, variables, statistics, layersdepths, localPath = getwd())
}
\arguments{
\item{locations}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION}

\item{statistics}{PARAM_DESCRIPTION}

\item{layersdepths}{PARAM_DESCRIPTION}

\item{localPath}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{gather}},\code{\link[tidyr]{pivot_wider}},\code{\link[tidyr]{nest}},\code{\link[tidyr]{c("nest", "nest", "nest")}},\code{\link[tidyr]{separate}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}},\code{\link[dplyr]{select}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{ranking}},\code{\link[dplyr]{mutate-joins}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
}
