% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPolaris-package.R
\docType{package}
\name{XPolaris-package}
\alias{XPolaris}
\alias{XPolaris-package}
\title{XPolaris: Retrieving Soil Data from POLARIS}
\description{
The POLARIS database <http://hydrology.cee.duke.edu/POLARIS/> 
    is a 30-meter probabilistic soil series map of the contiguous United 
    States (US). It represents an optimization of the Soil Survey 
    Geographic database. This R package facilitates the access 
    to large amounts of soil data within the US, currently 
    stored online as raster images (TIFF).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmrosso/XPolaris}
  \item Report bugs at \url{https://github.com/lhmrosso/XPolaris/issues}
}

}
\author{
\strong{Maintainer}: Luiz Moro Rosso \email{lhmrosso@ksu.edu}

Authors:
\itemize{
  \item Andre Borja Reis \email{andrefbr@ksu.edu}
  \item Adrian Correndo \email{correndo@ksu.edu}
  \item Ignacio Ciampitti \email{ciampitti@ksu.edu}
}

}
\keyword{internal}
