% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsoil.R
\name{xsoil}
\alias{xsoil}
\title{xsoil}
\usage{
xsoil(ximages_output, localPath = getwd())
}
\arguments{
\item{ximages_output}{The \code{data.frame} output from \code{ximages} function.}

\item{localPath}{Path in the user's machine to store the images. Default: getwd()}
}
\value{
It returns a \code{data.frame} object and exports a \code{.csv} file with the soil data. The \code{.csv} file will be save under the \code{POLARISOut} directory.
}
\description{
Extracting soil data from downloaded raster images.
}
\details{
This function must be executed after downloading the images (excuting the \code{ximages} function).
}
\examples{
\dontrun{

# df_test <- exkansas
#
# df_ximages <- ximages(locations = df_test,
#                       variables = c('clay'),
#                       statistics = c('mean'),
#                       layersdepths = c('0_5'))
#
# df_xsoil <- ximages(locations = df_ximages)
}
}
\seealso{
\code{\link[XPolaris]{xplot}},\code{\link[XPolaris]{ximages}}
}
